fix WEBUI_Picking_ReturnQtyToSourceHU

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/picking/pickingslot/process/WEBUI_Picking_ReturnQtyToSourceHU.java
@@ -1,24 +1,34 @@
 package de.metas.ui.web.picking.pickingslot.process;
 
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_CU;
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_HU;
+import com.google.common.collect.ImmutableSet;
+import de.metas.handlingunits.HuId;
+import de.metas.handlingunits.IHandlingUnitsBL;
+import de.metas.handlingunits.model.I_M_HU;
+import de.metas.handlingunits.picking.PickingCandidateService;
+import de.metas.handlingunits.picking.requests.RemoveQtyFromHURequest;
+import de.metas.handlingunits.sourcehu.HuId2SourceHUsService;
+import de.metas.handlingunits.storage.IHUStorage;
+import de.metas.handlingunits.storage.IHUStorageFactory;
+import de.metas.handlingunits.storage.IProductStorage;
+import de.metas.process.IProcessPrecondition;
+import de.metas.process.Param;
+import de.metas.process.ProcessPreconditionsResolution;
+import de.metas.product.ProductId;
+import de.metas.ui.web.handlingunits.HUEditorRowType;
+import de.metas.ui.web.picking.pickingslot.PickingSlotRow;
+import de.metas.util.Services;
+import lombok.NonNull;
+import org.adempiere.exceptions.AdempiereException;
+import org.adempiere.exceptions.FillMandatoryException;
+import org.compiere.SpringContextHolder;
 
 import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Objects;
 
-import org.springframework.beans.factory.annotation.Autowired;
-
-import de.metas.handlingunits.HuId;
-import de.metas.handlingunits.picking.PickingCandidateService;
-import de.metas.handlingunits.picking.requests.RemoveQtyFromHURequest;
-import de.metas.handlingunits.sourcehu.HuId2SourceHUsService;
-import de.metas.process.IProcessPrecondition;
-import de.metas.process.Param;
-import de.metas.process.ProcessPreconditionsResolution;
-import de.metas.ui.web.handlingunits.HUEditorRowType;
-import de.metas.ui.web.picking.pickingslot.PickingSlotRow;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_CU;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_HU;
 
 /*
  * #%L
@@ -46,11 +56,9 @@ public class WEBUI_Picking_ReturnQtyToSourceHU
 		extends WEBUI_Picking_With_M_Source_HU_Base
 		implements IProcessPrecondition
 {
-	@Autowired
-	private PickingCandidateService pickingCandidateService;
-
-	@Autowired
-	private HuId2SourceHUsService sourceHUsRepository;
+	private final PickingCandidateService pickingCandidateService = SpringContextHolder.instance.getBean(PickingCandidateService.class);
+	private final HuId2SourceHUsService sourceHUsRepository = SpringContextHolder.instance.getBean(HuId2SourceHUsService.class);
+	private final IHandlingUnitsBL handlingUnitsBL = Services.get(IHandlingUnitsBL.class);
 
 	private static final String PARAM_QTY_CU = "QtyCU";
 	@Param(parameterName = PARAM_QTY_CU, mandatory = true)
@@ -86,14 +94,20 @@ public class WEBUI_Picking_ReturnQtyToSourceHU
 	}
 
 	@Override
-	protected String doIt() throws Exception
+	protected String doIt()
 	{
+		if (qtyCU == null || qtyCU.signum() <= 0)
+		{
+			throw new FillMandatoryException("QtyCU");
+		}
+
 		final PickingSlotRow pickingSlotRow = getSingleSelectedRow();
+		final HuId huId = pickingSlotRow.getHuId();
 
 		pickingCandidateService.removeQtyFromHU(RemoveQtyFromHURequest.builder()
 				.qtyCU(qtyCU)
-				.huId(pickingSlotRow.getHuId())
-				.productId(pickingSlotRow.getHuProductId())
+				.huId(huId)
+				.productId(getProductId(huId))
 				.build());
 
 		invalidateView();
@@ -107,5 +121,29 @@ public class WEBUI_Picking_ReturnQtyToSourceHU
 		final HuId huId = getSingleSelectedRow().getHuId();
 		final Collection<HuId> sourceHUs = sourceHUsRepository.retrieveMatchingSourceHUIds(huId);
 		return !sourceHUs.isEmpty();
+	}
+
+	private ProductId getProductId(@NonNull final HuId huId)
+	{
+		final I_M_HU hu = handlingUnitsBL.getById(huId);
+		final IHUStorageFactory storageFactory = handlingUnitsBL.getStorageFactory();
+		final IHUStorage storage = storageFactory.getStorage(hu);
+
+		final ImmutableSet<ProductId> productIds = storage.getProductStorages().stream()
+				.map(IProductStorage::getProductId)
+				.collect(ImmutableSet.toImmutableSet());
+
+		if (productIds.isEmpty())
+		{
+			throw new AdempiereException("Selected HU is empty");
+		}
+		else if (productIds.size() == 1)
+		{
+			return productIds.iterator().next();
+		}
+		else
+		{
+			throw new AdempiereException("Selected HU has more than one product");
+		}
 	}
 }
